updated for mco3
it is not possible for a user to reserve a seat on a timeslot that had already been previously taken.

make it such that for the technician walk-in reservation, the technician must enter the student's email instead of their name. This allows the reservation to still be linked to the walk-in student. This option should also verify whether the entered email is valid or not. It should also still be possible for the reservation to be made anonymously."

ooking for availability can be improved. Try providing 2 separate dropdowns (starttime and endtime), which would then display the availability of the seats in the lab for the entire specified timespan. You can also take advantage of colors to represent a seat's availability e.g., transparent = free, orange = occupied, but have free timeslots, red = fully occupied